+++ src/components/index.js
@@ -17,3 +17,4 @@ export {default as AuthSuccess} from './auth/authSuccess';
 export {default as VerifyEmail} from './auth/forms/VerifyEmail';
 export {default as NewPassword} from './auth/forms/NewPassword'
 export {default as PageNotFoundComponent} from './pageNotFound';
+export {default as EmptyOrders} from './orders/empty-orders';
