+++ src/components/overview/index.js
@@ -1,14 +1,23 @@
-import React from "react";
+import React, { useEffect } from "react";
 import Welcome from "../dashboard/welcome";
 import Dashboard from "../dashboard/dashboard";
-import { OVERVIEW_DATA } from "../../data";
+import { useDispatch, useSelector } from "react-redux";
+import { getOrdersData, resetStore } from "../../redux/action";
 
 
 const Overview = () => {
+    const dispatch = useDispatch();
 
+    useEffect(() => {
+        dispatch(resetStore())
+        dispatch(getOrdersData(1))
+    }, [])
+
+    const ordersData = [useSelector((state) => state.ordersData)];
+    const orders = []
     return (
         <div >
-            {OVERVIEW_DATA.length < 0 ?
+            {orders.length === 0 ?
                 <Welcome />
                 :
                 <Dashboard />
